Update with option-based approach
It turns out that working with `derangements (option \alpha)` was much
easier, mostly because working with `\alpha` is easier than working with
the set `{a}^c` for some `a`.

So this got around a lot of the issues with sets/types and giving things
names to appease the elaborator. There's still a few more TODOs, but
I've incorporated the changes that were suggested (e.g. no non-terminal
simps, style guidelines, pull things into files).