Remove `Prop` parameters in `linear_independent.extend`
The previous version took two proofs as parameters, which we can get rid of
and get a more general construction that still has the desired properties
if the previous assumptions hold. (In fact, linear independence does not
even require the other assumption!)

The phrasing of `exists_extend` is still quite messy, so before PR'ing I want
to clean it up.