chore(field_theory): turn `intermediate_field.subalgebra.equiv_of_eq` into `intermediate_field.equiv_of_eq`
There was a definition of `subalgebra.equiv_of_eq` in the `intermediate_field`
namespace which is equal to the original `subalgebra.equiv_of_eq` definition.
Meanwhile, there was no `intermediate_field.equiv_of_eq`. So I decided to turn
the duplicate and weirdly named lemma into a new useful lemma.