feat(data/list/nodup_permutations): nodup_permutations
Using a cardinality argument, prove that the list of permutations of a list
is nodup if the list is nodup.
TODO: prove the reverse implication.
Implementation: provide an explicit equiv between
the perm Prop and the equiv over indices.
TODO: prove that this equiv moves indices as expected
TODO: generalize to sublists/embeddings