feat(linear_algebra): Smith normal form for submodules over a PID
This PR expands on `submodule.basis_of_pid` by showing that this basis can be
chosen in "Smith normal form". That is: if `M` has a basis and `N ≤ M`, then we
can choose a basis `bN` for `N` and `bM` for `M`, such that the inclusion map
from `N` to `M` expressed in these bases is a diagonal matrix in Smith normal
form.