feat(data/list/rotate): cyclic_permutations (#8678)
For `l ~ l'` we have `list.permutations`. We provide the list of cyclic permutations of `l` such that all members are `l ~r l'`. This relationship is proven, as well as the induced `nodup` of the list of cyclic permutants.

This also simplifies the `cycle.list` definition, and removed the requirement for decidable equality in it.