feat(linear_algebra/bilinear_form): the dual basis for a nondegenerate bilinear form
Let `B` be a nondegenerate bilinear form on a finite-dimensional vector space `V` and `b` a basis on `V`. Then there is a "dual basis" `d` such that `d.repr x i = B x (b i)`.

In a follow-up PR, we'll show that the trace form for `L / K` produces a dual basis consisting only of elements integral over the ring of integers of `K`.