feat(*): remove the `fintype` requirement from matrices. (#8810)
For historical reasons, `matrix` has always had `fintype` attached to it. I remove this artificial limitation, but with a big caveat; multiplication is currently defined in terms of the dot product, which requires finiteness; therefore, any multiplicative structure on matrices currently requires fintypes. This can be removed in future contributions, however.



Co-authored-by: Eric <37984851+ericrbg@users.noreply.github.com>